requirements.txt
pytest
pyyaml
requests